Qt5 QDateTime: Modify QDateTimeParser to new calendar and format api
Modify the date time parser to initially support the minimum changes
required for calendar systems and full CLDR format codes.  Note many
codes are not valid for parsing, such as quarters or narrow format.

Note that this parser will be replaced in the future with a simpler
QLocale specific parser and this parser will be moved into QDateTimeEdit
directly.

As always, use the QT4_COMAT directive to use the old Qt4 format codes.