QLocale: Add CLDR/ICU style date/time format/parse API
This is an initial push for api review purposes.

This change implements an new date/time parse/format API in line with
the CLDR/ICU standards.  This is in anticipation of a switch to ICU in
Qt 5.1.

This patch is the minimum required and translates some format codes to
the current backend.  Further patches will improve the individual
system implementations.

This change has not yet been compiled on Mac or Windows, but will be.

Change-Id: I0fbac1d6d4dad9f6681f4f77097f99fca68a98b5