SIC: QLocale - Change number symbols API from QChar to QString
To conform with CLDR / ICU, make the number symbol API QString instead
of QChar. In 5.0 we will still only have data with a single char, but
this prepares for 5.1 when using ICU will include string symbols.  Note
that the new code in QTextStream cannot be explicitly tested for this
reason.

Change-Id: Ic0d84a6f0151b74e01128318c2e452edce9c8129