QDateTime - Improve and expose Qt::OffsetFromUtc
The Qt::OffsetFromUtc TimeSpec was introduced in Qt 4.4 but the methods
were never made public in the apidox effectively meaning the feature
was never used.  It was also incomplete and inconsitent in a few places.

This change cleans up the implementation and exports new public API for
using the TimeSpec using new method names consistent with the new
QTimeZone support.

The behavior of one constructor has changed slightly to be consistent
with the rest of the feature, but as this behavior was never
documented this should be acceptable.

Task-number: QTBUG-26161
Task-number: QTBUG-29666

Change-Id: If03899ef78dde6030aa0a55ac80e26a2067f0106