QDateTime - Change to use QLocale formatter
Clean-up the implementation of toString() and fromString() methods in
QDate, QTime and QDateTime and switch them to the QLocale formatter.

1) Switch all methods to call the QLocale methods instead of the
   QDateTime formatter, except for the ISODate and TextDate.
2) Remove the now redundant QDateTime formatter.
3) Clean-up the ISODate and TextDate code to be more consistent in
   behavior, especially when handling TimeSpec.
4) Add an abbreviation() call to QDateTime to return the time zone
   abbreviation.
5) Clean-up the QLocale formatter and fix its time zone handling.
6) Move most formatter test cases to QLocale and fix some test cases
   for consistency.
7) Improve the documentation where it is ambiguous or wrong.

This changes some corner-case behavior in TextDate and ISODate, but
this either fixes bugs or makes the behavior match the documentation.

Some reformatting of code is included where the method was being
changed anyway.

Change-Id: Id848524346cf932a5243240ca56a1b100fa0ef9f