WIP - Proposal to add QLocale::NoCountry
This a proof-of-concept for adding NoCountry to the QLocale::Country
enum.  In QTimeZone a time zone usually has a country it applies to,
e.g. Europe/Berlin applies to Germany. However some time zones such as
UTC do not have a country. In the api using QLocale::AnyCountry for
these zones poses a problem as AnyCountry is probably better used to
return any country that matches.  Adding QLocale::NoCountry makes the
QTimeZone api simpler and clearer.

However, I'm reluctant to do this as it will have consequences for
anyone using QLocale::Country to suddenly find a non-country in the
middle of their pick-lists, especially as like AnyCountry it will
return a null name.  Also in QLocale it could be taken to mean a
locale with no country component, e.g. "en" instead of "en_US". One
option is to have NoCountry with an enum value of -1 but I'm not
sure how portable or safe that is in public api.

Change-Id: Ia10c003da40413f083d34bb52790195faf25b03d