WIP - QDateTime - Add proper occurrence support
Add proper support to QDateTime for occurrences in LocalTime and
TimeZone.  First and Second occurrences of a time happen in the
hour after the change from Daylight Time to Standard Time, i.e.
when the clock is turned back one hour from 3am to 2am we have two
occurrences of 2am local time which we need to tell apart.

This change removes the use of the QDateTimePrivate::Spec which was
intended to allow for the tracking of occurrences but never worked
properly as there was no way to initially set the second occurrence.
It also simplifies the QTimeZone code to come which would otherwise
have required 3 new values.  Instead the real Qt::TimeSpec is
stored and a new enum tracks the occurrence separately.  This does
increase the memory usage but greatly simplifies the code and
improves performance.

Because the changes affected the signatures of several private methods
the opportunity was taken to improve their readability by changing to
pass-by-pointer where the methods modify the parameters.

Change-Id: I26bec78a58b8f71b5cfcda3ad713dc646f3727e8