QDateTime - Extend parser/widget supported dates
Slightly extend the range of dates supported by the parser and widgets.

Extending the range further requires extensive changes to the parser
and widgets, in particular removing the 4 digit positive year
restriction in the parser.

[ChangeLog][Important Behavior Changes] The supported and default date
range for some widgets has changed.  QDateTimeEdit now supports and
defaults to 100 AD to 9999 AD.  QCalendarWidget now defaults to 9999 BC
to 9999 AD. Date parsing in QLocale and QDateTime now supports a range
of 1 AD to 9999 AD.

Task-number: QTBUG-31272

Change-Id: I68eb4a6b6bca3b021306faf225e6aad37ea6520a