QPrintEngine - Fix PPK_CollateCopies
Mac supports Collate Copies using native api, so add support.  Note
this is mostly only useful for setting the print dialog default, as
Mac supports server-side multiple copies so the app will never need to
collate the copies itself.

Change PDF and Windows to default to collate true to match Mac as this
is the behaviour users expect.

Task-number: QTBUG-27724

Change-Id: Ia43dbc260b3a71aa5b267cca54c168ffbea794fc